Renamed project, added new empty form
