+++ MyLibrarian.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28010.2026
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Librius", "Librius.csproj", "{B4499854-E465-432B-9741-90BEDA510351}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MyLibrarian", "MyLibrarian.csproj", "{B4499854-E465-432B-9741-90BEDA510351}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
